Add files via upload
Making sure to push to the right directory this time...

Added a custom titlecase function to models in an attempt to catch and correct user errors when inputting titles.  It should properly capitalize *most* titles, although there are still a couple bugs to work out (see comments in the function).

Assigned a default value and possible accepted values to shelf (models.py) in order to make views.py's auto-assign shelf feature work as intended.  A book with shelf "0" in the database is an indicator of a bug.

On that note, added a feature to views.py to auto-assign a new book to the proper shelf based on genre.  However, this only assigns to shelves xA and not xB and xC, so a system for assigning sub-shelves would be a good next step.

Submitting a valid form to new_book_form renders the new_book_confirmation page and tells the user on which shelf to put the new book.  From there the user can return to a blank new_book_form to catalog another book.  (This may have made the new_book_confirmation function in views.py obsolete -- stay tuned for updates on that.)

Overall a lot of new features added and a lot of new bugs to work on!